Create method get, post and put

+++ backend/src/main/java/com/jonas/backend/repositories/EnderecoRepository.java
@@ -3,6 +3,6 @@ package com.jonas.backend.repositories;
 import com.jonas.backend.entities.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoRepository extends JpaRepository<Endereco, Integer>{
+public interface EnderecoRepository extends JpaRepository<Endereco, Long>{
     
 }
